core: Do not send client metadata before connecting to room
