+++ .changeset/nervous-shrimps-deny.md
@@ -1,0 +1,5 @@
+---
+"@whereby.com/core": minor
+---
+
+Do not send client metadata before connecting to room
