+++ packages/core/src/redux/index.ts
@@ -17,6 +17,7 @@ export * from "./slices/organization";
 export * from "./slices/remoteParticipants";
 export * from "./slices/room";
 export * from "./slices/roomConnection";
+export * from "./slices/roomConnection/selectors";
 export * from "./slices/rtcAnalytics";
 export * from "./slices/spotlights";
 export * from "./slices/streaming";
